feat: add dedup checks
ensure the user is not accidentally uploading the same input files
or enqueueing the same job multiple times on subsequent client runs. the
scope for this is batch specific. once the batch of jobs is finished and
the client side memory is wiped the user can upload and enqueue the same
jobs if they want to